blackjack_web/requirements.txt
Flask
google-generativeai
python-dotenv 